[[ DefaultHandlers ]] Improve default handler construction This patch reinstates the use of default scripts specified by widgets and allows them to have multiple handlers within them.
It also adds the ability to override default scripts for handlers
dependent on 'classic' object type, for example different default
scripts for a standard button and an option menu.

Finally, if a default script override is not present, info from the
documentation is used to construct a default handler with params
and summary information.