[Bugfix-20140] Gutter better Script Editor scrolling
Drag Scrolling and arrow keys appear to be “blocking” messages. As a result, gutter updates are delayed when “send message in time” is used. This results in gutter redraws lagging behind the script editor scroll.

When up/down arrows are pressed, added a wait for 0 ms to allow handling of Gutter update while arrow keys are held down.