[bugfix-201140] Removed Mouseup send in
Found when clicking gutter in succession, mouseup messages were not handled correctly. Removed the send in 0 milliseconds to handle mouseUp. Prior comment mentions that the send in time is needed incase we delete a breakpoint. Reviewing the code I can not see why send in time matters when breakpoints are deleted.